+++ src/main/java/com/example/tictactoe/UI_Window.java
@@ -1,10 +1,11 @@
 package com.example.tictactoe;
 
-public class UI_field {
+public class UI_Window {
     public static final int WINDOW_WIDTH = 300;
     public static final int WINDOW_HEIGHT = 400;
 
     public static final int TITLE_HEIGHT = 100;
     public static final int GAME_HEIGHT = 300;
+    public static final int SIGNIN_HEIGHT = 300;
 
 }
